phydynR: Model-based coalescent simulation and likelihood for phylodynamic inference

This package provides functions for defining infectious disease models or ecological process models in terms of ODEs (ordinary differential equations) or SDEs (stochastic differential equations); functions for simulating trees conditional on process models; and functions for computing likelihoods of trees conditional on a process model.

References

1. Volz, Erik M. Complex population dynamics and the coalescent under neutrality. Genetics 190.1 (2012): 187-201.

2. Volz, Erik M., Katia Koelle, and Trevor Bedford. Viral phylodynamics. PLoS Comput Biol 9.3 (2013): e1002947.

See also

Author

Maintainer: Erik Volz erik@erikvolz.info (ORCID)

Other contributors:

Fabricia F. Nascimento thednainus@yahoo.com (ORCID) [contributor]